backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]==0.24.0
google-genai>=1.0.0
anthropic>=1.0.0
python-dotenv==1.0.0
redis==5.0.1
python-multipart==0.0.6
